requirements.txt
psycopg2==2.7.3.2
Django==2.0
django-test-plus==1.0.21
langsplit==0.1.6
bigchaindb-driver==0.4.1
djangorestframework==3.7.7
django-rest-auth==0.9.2
django-cors-headers==2.1.0
coreapi==2.3.3
django-simple-captcha==0.5.6
django-filter==1.1.0
django-allauth==0.34.0
argon2==0.1.10
argon2-cffi==16.3.0
